Capstone Project Week 4 INTRODUCTION
# Introduction / Business Problem

The aim of this project is to find suitable locations to open a hotel in karol bagh greater metropolitan area [Karol bagh  and its suburbns],Delhi ( India ).

The first requirement is that the new hotel should be easily accessible by its prospective customers and more specifically it should be located near a metro station. The number of hotels already existing in an area should also be considered so that fierce competition be avoided if possible.

Apart from the obvious intended stakeholders, entrepreneurs looking to start a hotel business, similar methodology could be used for other specific types of businesses. It can serve as an initial starting point of locations to consider to start their business.

For the project objectives to be achieved, python geolocation libraries will be used, along with the Foursquare API. Also, in order to create clusters of similar candidate locations, the K-Means machine learning clustering algorithm will be used.